refactor(kyosei): エントリポイントをsrc/binに移動し、submit-reviewモジュールを分割
`bin/`ディレクトリのエントリポイントを`src/bin/`に移動しました。
Claude Code Pluginが`bin/`にPATHを通す仕様があるため、
トランスパイル前提の`.ts`ファイルが直接実行される混乱を避けるためです。

あわせて`src/submit-review.ts`の命名がCLIエントリポイント`src/bin/submit-review.ts`と
重複して紛らわしかったため、責務ごとに3つのモジュールに分割しました。

- `src/review-comment-body.ts`: GitHub Alert形式とタグラベルのコメント本文フォーマット
- `src/review-decoder.ts`: JSON入力のデコード・バリデーション(`decodeReviewSubmission`)
- `src/review-poster.ts`: GitHub APIへの投稿(`submitReview`、`previewReview`)
